+++ src/vite-env.d.ts
@@ -2,7 +2,7 @@
 /// <reference types="vite-plugin-pwa/client" />
 
 interface ImportMetaEnv {
-  readonly VITE_GOOGLE_CLIENT_ID: string;
+  readonly VITE_LEALTEK_API_URL: string;
 }
 
 interface ImportMeta {